+++ connect/src/main/java/com/flip/connect/data/model/UpdateModel.java
@@ -1,0 +1,8 @@
+package com.flip.connect.data.model;
+
+/**
+ * Created by ltorres on 19/07/2017.
+ */
+
+public class UpdateModel {
+}
