+++ connect/src/main/java/com/flip/connect/domain/model/BaseResponse.java
@@ -1,0 +1,9 @@
+package com.flip.connect.data.model;
+
+/**
+ * Created by ltorres on 19/07/2017.
+ */
+
+public class BaseResponse {
+    private 
+}
